More work on fixing     public void testStackdriverErrorHandling()  unit test
