+++ .gitignore
@@ -6,3 +6,4 @@ target/
 *.swp
 /bin/
 *.iml
+.temp-*
